# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Freelance CI API",
        "description": "API REST de la plateforme Freelance CI — connexion freelances, clients et entrepreneurs en Côte d'Ivoire. Paiements via Genius Pay (Mobile Money, Carte), authentification Supabase JWT.",
        "contact": {
            "name": "Freelance CI",
            "email": "contact@freelance-ci.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.freelance-ci.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/projects/{project}/quotes": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "List quotes for a project",
                "operationId": "c3fff716684df2606d318a45ecd46d8a",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quotes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Quote"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Submit a quote for a project",
                "operationId": "b7e91c34149e5a57575ce7d0f354a508",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "duration_unit": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quote submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Quote"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to create quote",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/quotes/{id}": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Get a quote by ID",
                "operationId": "451bbe89f2be1f8ddfa73d76dbb0f227",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quote ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Quote"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quote not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Update a quote",
                "operationId": "a308c86572f538405c2da9a796ab1a59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quote ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "duration_unit": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quote updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Quote"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Quote not found or cannot be updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Withdraw a quote",
                "operationId": "e0083f49ae2de0d7945c879c337475b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quote ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote withdrawn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quote not found or unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/quotes/{id}/accept": {
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Accept a quote and create a contract",
                "operationId": "083497648db9203dcf9b4251b9453663",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quote ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote accepted, contract created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Contract"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to accept quote",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/quotes/{id}/refuse": {
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Refuse a quote",
                "operationId": "369d5e1b20c6efab51471a6e02768cdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quote ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to refuse quote",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Quote": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "estimated_duration": {
                        "type": "integer",
                        "nullable": true
                    },
                    "proposal": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "refused",
                            "withdrawn"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Contract": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "platform_fee": {
                        "type": "number",
                        "format": "float"
                    },
                    "freelance_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "completed",
                            "cancelled",
                            "disputed"
                        ]
                    },
                    "client_signed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "freelance_signed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "JWT Supabase — coller le token reçu après connexion",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}